+++ sdk/datadog/azure-resourcemanager-datadog/src/main/java/com/azure/resourcemanager/datadog/MicrosoftDatadogManager.java
@@ -24,6 +24,7 @@ import com.azure.core.management.profile.AzureProfile;
 import com.azure.core.util.Configuration;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.resourcemanager.datadog.fluent.MicrosoftDatadogClient;
+import com.azure.resourcemanager.datadog.implementation.BillingInfoesImpl;
 import com.azure.resourcemanager.datadog.implementation.CreationSupportedsImpl;
 import com.azure.resourcemanager.datadog.implementation.MarketplaceAgreementsImpl;
 import com.azure.resourcemanager.datadog.implementation.MicrosoftDatadogClientBuilder;
@@ -32,6 +33,7 @@ import com.azure.resourcemanager.datadog.implementation.MonitorsImpl;
 import com.azure.resourcemanager.datadog.implementation.OperationsImpl;
 import com.azure.resourcemanager.datadog.implementation.SingleSignOnConfigurationsImpl;
 import com.azure.resourcemanager.datadog.implementation.TagRulesImpl;
+import com.azure.resourcemanager.datadog.models.BillingInfoes;
 import com.azure.resourcemanager.datadog.models.CreationSupporteds;
 import com.azure.resourcemanager.datadog.models.MarketplaceAgreements;
 import com.azure.resourcemanager.datadog.models.MonitoredSubscriptions;
@@ -46,7 +48,9 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
-/** Entry point to MicrosoftDatadogManager. */
+/**
+ * Entry point to MicrosoftDatadogManager.
+ */
 public final class MicrosoftDatadogManager {
     private MarketplaceAgreements marketplaceAgreements;
 
@@ -55,6 +59,8 @@ public final class MicrosoftDatadogManager {
     private Monitors monitors;
 
     private Operations operations;
+
+    private BillingInfoes billingInfoes;
 
     private TagRules tagRules;
 
@@ -67,18 +73,16 @@ public final class MicrosoftDatadogManager {
     private MicrosoftDatadogManager(HttpPipeline httpPipeline, AzureProfile profile, Duration defaultPollInterval) {
         Objects.requireNonNull(httpPipeline, "'httpPipeline' cannot be null.");
         Objects.requireNonNull(profile, "'profile' cannot be null.");
-        this.clientObject =
-            new MicrosoftDatadogClientBuilder()
-                .pipeline(httpPipeline)
-                .endpoint(profile.getEnvironment().getResourceManagerEndpoint())
-                .subscriptionId(profile.getSubscriptionId())
-                .defaultPollInterval(defaultPollInterval)
-                .buildClient();
+        this.clientObject = new MicrosoftDatadogClientBuilder().pipeline(httpPipeline)
+            .endpoint(profile.getEnvironment().getResourceManagerEndpoint())
+            .subscriptionId(profile.getSubscriptionId())
+            .defaultPollInterval(defaultPollInterval)
+            .buildClient();
     }
 
     /**
      * Creates an instance of MicrosoftDatadog service API entry point.
-     *
+     * 
      * @param credential the credential to use.
      * @param profile the Azure profile for client.
      * @return the MicrosoftDatadog service API instance.
@@ -91,7 +95,7 @@ public final class MicrosoftDatadogManager {
 
     /**
      * Creates an instance of MicrosoftDatadog service API entry point.
-     *
+     * 
      * @param httpPipeline the {@link HttpPipeline} configured with Azure authentication credential.
      * @param profile the Azure profile for client.
      * @return the MicrosoftDatadog service API instance.
@@ -104,14 +108,16 @@ public final class MicrosoftDatadogManager {
 
     /**
      * Gets a Configurable instance that can be used to create MicrosoftDatadogManager with optional configuration.
-     *
+     * 
      * @return the Configurable instance allowing configurations.
      */
     public static Configurable configure() {
         return new MicrosoftDatadogManager.Configurable();
     }
 
-    /** The Configurable allowing configurations to be set. */
+    /**
+     * The Configurable allowing configurations to be set.
+     */
     public static final class Configurable {
         private static final ClientLogger LOGGER = new ClientLogger(Configurable.class);
 
@@ -183,8 +189,8 @@ public final class MicrosoftDatadogManager {
 
         /**
          * Sets the retry options for the HTTP pipeline retry policy.
-         *
-         * <p>This setting has no effect, if retry policy is set via {@link #withRetryPolicy(RetryPolicy)}.
+         * <p>
+         * This setting has no effect, if retry policy is set via {@link #withRetryPolicy(RetryPolicy)}.
          *
          * @param retryOptions the retry options for the HTTP pipeline retry policy.
          * @return the configurable object itself.
@@ -201,8 +207,8 @@ public final class MicrosoftDatadogManager {
          * @return the configurable object itself.
          */
         public Configurable withDefaultPollInterval(Duration defaultPollInterval) {
-            this.defaultPollInterval =
-                Objects.requireNonNull(defaultPollInterval, "'defaultPollInterval' cannot be null.");
+            this.defaultPollInterval
+                = Objects.requireNonNull(defaultPollInterval, "'defaultPollInterval' cannot be null.");
             if (this.defaultPollInterval.isNegative()) {
                 throw LOGGER
                     .logExceptionAsError(new IllegalArgumentException("'defaultPollInterval' cannot be negative"));
@@ -222,15 +228,13 @@ public final class MicrosoftDatadogManager {
             Objects.requireNonNull(profile, "'profile' cannot be null.");
 
             StringBuilder userAgentBuilder = new StringBuilder();
-            userAgentBuilder
-                .append("azsdk-java")
+            userAgentBuilder.append("azsdk-java")
                 .append("-")
                 .append("com.azure.resourcemanager.datadog")
                 .append("/")
-                .append("1.0.0");
+                .append("1.0.0-beta.1");
             if (!Configuration.getGlobalConfiguration().get("AZURE_TELEMETRY_DISABLED", false)) {
-                userAgentBuilder
-                    .append(" (")
+                userAgentBuilder.append(" (")
                     .append(Configuration.getGlobalConfiguration().get("java.version"))
                     .append("; ")
                     .append(Configuration.getGlobalConfiguration().get("os.name"))
@@ -255,38 +259,28 @@ public final class MicrosoftDatadogManager {
             policies.add(new UserAgentPolicy(userAgentBuilder.toString()));
             policies.add(new AddHeadersFromContextPolicy());
             policies.add(new RequestIdPolicy());
-            policies
-                .addAll(
-                    this
-                        .policies
-                        .stream()
-                        .filter(p -> p.getPipelinePosition() == HttpPipelinePosition.PER_CALL)
-                        .collect(Collectors.toList()));
+            policies.addAll(this.policies.stream()
+                .filter(p -> p.getPipelinePosition() == HttpPipelinePosition.PER_CALL)
+                .collect(Collectors.toList()));
             HttpPolicyProviders.addBeforeRetryPolicies(policies);
             policies.add(retryPolicy);
             policies.add(new AddDatePolicy());
             policies.add(new ArmChallengeAuthenticationPolicy(credential, scopes.toArray(new String[0])));
-            policies
-                .addAll(
-                    this
-                        .policies
-                        .stream()
-                        .filter(p -> p.getPipelinePosition() == HttpPipelinePosition.PER_RETRY)
-                        .collect(Collectors.toList()));
+            policies.addAll(this.policies.stream()
+                .filter(p -> p.getPipelinePosition() == HttpPipelinePosition.PER_RETRY)
+                .collect(Collectors.toList()));
             HttpPolicyProviders.addAfterRetryPolicies(policies);
             policies.add(new HttpLoggingPolicy(httpLogOptions));
-            HttpPipeline httpPipeline =
-                new HttpPipelineBuilder()
-                    .httpClient(httpClient)
-                    .policies(policies.toArray(new HttpPipelinePolicy[0]))
-                    .build();
+            HttpPipeline httpPipeline = new HttpPipelineBuilder().httpClient(httpClient)
+                .policies(policies.toArray(new HttpPipelinePolicy[0]))
+                .build();
             return new MicrosoftDatadogManager(httpPipeline, profile, defaultPollInterval);
         }
     }
 
     /**
      * Gets the resource collection API of MarketplaceAgreements.
-     *
+     * 
      * @return Resource collection API of MarketplaceAgreements.
      */
     public MarketplaceAgreements marketplaceAgreements() {
@@ -298,7 +292,7 @@ public final class MicrosoftDatadogManager {
 
     /**
      * Gets the resource collection API of CreationSupporteds.
-     *
+     * 
      * @return Resource collection API of CreationSupporteds.
      */
     public CreationSupporteds creationSupporteds() {
@@ -310,7 +304,7 @@ public final class MicrosoftDatadogManager {
 
     /**
      * Gets the resource collection API of Monitors. It manages DatadogMonitorResource.
-     *
+     * 
      * @return Resource collection API of Monitors.
      */
     public Monitors monitors() {
@@ -322,7 +316,7 @@ public final class MicrosoftDatadogManager {
 
     /**
      * Gets the resource collection API of Operations.
-     *
+     * 
      * @return Resource collection API of Operations.
      */
     public Operations operations() {
@@ -333,8 +327,20 @@ public final class MicrosoftDatadogManager {
     }
 
     /**
+     * Gets the resource collection API of BillingInfoes.
+     * 
+     * @return Resource collection API of BillingInfoes.
+     */
+    public BillingInfoes billingInfoes() {
+        if (this.billingInfoes == null) {
+            this.billingInfoes = new BillingInfoesImpl(clientObject.getBillingInfoes(), this);
+        }
+        return billingInfoes;
+    }
+
+    /**
      * Gets the resource collection API of TagRules. It manages MonitoringTagRules.
-     *
+     * 
      * @return Resource collection API of TagRules.
      */
     public TagRules tagRules() {
@@ -346,26 +352,26 @@ public final class MicrosoftDatadogManager {
 
     /**
      * Gets the resource collection API of SingleSignOnConfigurations. It manages DatadogSingleSignOnResource.
-     *
+     * 
      * @return Resource collection API of SingleSignOnConfigurations.
      */
     public SingleSignOnConfigurations singleSignOnConfigurations() {
         if (this.singleSignOnConfigurations == null) {
-            this.singleSignOnConfigurations =
-                new SingleSignOnConfigurationsImpl(clientObject.getSingleSignOnConfigurations(), this);
+            this.singleSignOnConfigurations
+                = new SingleSignOnConfigurationsImpl(clientObject.getSingleSignOnConfigurations(), this);
         }
         return singleSignOnConfigurations;
     }
 
     /**
      * Gets the resource collection API of MonitoredSubscriptions. It manages MonitoredSubscriptionProperties.
-     *
+     * 
      * @return Resource collection API of MonitoredSubscriptions.
      */
     public MonitoredSubscriptions monitoredSubscriptions() {
         if (this.monitoredSubscriptions == null) {
-            this.monitoredSubscriptions =
-                new MonitoredSubscriptionsImpl(clientObject.getMonitoredSubscriptions(), this);
+            this.monitoredSubscriptions
+                = new MonitoredSubscriptionsImpl(clientObject.getMonitoredSubscriptions(), this);
         }
         return monitoredSubscriptions;
     }
@@ -373,7 +379,7 @@ public final class MicrosoftDatadogManager {
     /**
      * Gets wrapped service client MicrosoftDatadogClient providing direct access to the underlying auto-generated API
      * implementation, based on Azure REST API.
-     *
+     * 
      * @return Wrapped service client MicrosoftDatadogClient.
      */
     public MicrosoftDatadogClient serviceClient() {