+++ sdk/datadog/azure-resourcemanager-datadog/src/main/java/com/azure/resourcemanager/datadog/models/MonitoredSubscriptions.java
@@ -4,40 +4,43 @@
 
 package com.azure.resourcemanager.datadog.models;
 
-import com.azure.core.http.rest.PagedIterable;
 import com.azure.core.http.rest.Response;
 import com.azure.core.util.Context;
 
-/** Resource collection API of MonitoredSubscriptions. */
+/**
+ * Resource collection API of MonitoredSubscriptions.
+ */
 public interface MonitoredSubscriptions {
     /**
      * List the subscriptions currently being monitored by the Datadog monitor resource.
-     *
-     * @param resourceGroupName The name of the resource group. The name is case insensitive.
-     * @param monitorName Monitor resource name.
-     * @throws IllegalArgumentException thrown if parameters fail the validation.
-     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
-     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return the paginated response with {@link PagedIterable}.
-     */
-    PagedIterable<MonitoredSubscriptionProperties> list(String resourceGroupName, String monitorName);
-
-    /**
-     * List the subscriptions currently being monitored by the Datadog monitor resource.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param monitorName Monitor resource name.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
-     * @return the paginated response with {@link PagedIterable}.
+     * @return the request to update subscriptions needed to be monitored by the Datadog monitor resource along with
+     * {@link Response}.
      */
-    PagedIterable<MonitoredSubscriptionProperties> list(String resourceGroupName, String monitorName, Context context);
+    Response<MonitoredSubscriptionProperties> listWithResponse(String resourceGroupName, String monitorName,
+        Context context);
 
     /**
      * List the subscriptions currently being monitored by the Datadog monitor resource.
-     *
+     * 
+     * @param resourceGroupName The name of the resource group. The name is case insensitive.
+     * @param monitorName Monitor resource name.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return the request to update subscriptions needed to be monitored by the Datadog monitor resource.
+     */
+    MonitoredSubscriptionProperties list(String resourceGroupName, String monitorName);
+
+    /**
+     * List the subscriptions currently being monitored by the Datadog monitor resource.
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param monitorName Monitor resource name.
      * @param configurationName The configuration name. Only 'default' value is supported.
@@ -46,14 +49,14 @@ public interface MonitoredSubscriptions {
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return the request to update subscriptions needed to be monitored by the Datadog monitor resource along with
-     *     {@link Response}.
+     * {@link Response}.
      */
-    Response<MonitoredSubscriptionProperties> getWithResponse(
-        String resourceGroupName, String monitorName, String configurationName, Context context);
+    Response<MonitoredSubscriptionProperties> getWithResponse(String resourceGroupName, String monitorName,
+        String configurationName, Context context);
 
     /**
      * List the subscriptions currently being monitored by the Datadog monitor resource.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param monitorName Monitor resource name.
      * @param configurationName The configuration name. Only 'default' value is supported.
@@ -66,7 +69,7 @@ public interface MonitoredSubscriptions {
 
     /**
      * Updates the subscriptions that are being monitored by the Datadog monitor resource.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param monitorName Monitor resource name.
      * @param configurationName Configuration name.
@@ -78,7 +81,7 @@ public interface MonitoredSubscriptions {
 
     /**
      * Updates the subscriptions that are being monitored by the Datadog monitor resource.
-     *
+     * 
      * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param monitorName Monitor resource name.
      * @param configurationName Configuration name.
@@ -91,32 +94,32 @@ public interface MonitoredSubscriptions {
 
     /**
      * List the subscriptions currently being monitored by the Datadog monitor resource.
-     *
+     * 
      * @param id the resource ID.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return the request to update subscriptions needed to be monitored by the Datadog monitor resource along with
-     *     {@link Response}.
+     * {@link Response}.
      */
     MonitoredSubscriptionProperties getById(String id);
 
     /**
      * List the subscriptions currently being monitored by the Datadog monitor resource.
-     *
+     * 
      * @param id the resource ID.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return the request to update subscriptions needed to be monitored by the Datadog monitor resource along with
-     *     {@link Response}.
+     * {@link Response}.
      */
     Response<MonitoredSubscriptionProperties> getByIdWithResponse(String id, Context context);
 
     /**
      * Updates the subscriptions that are being monitored by the Datadog monitor resource.
-     *
+     * 
      * @param id the resource ID.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
@@ -126,7 +129,7 @@ public interface MonitoredSubscriptions {
 
     /**
      * Updates the subscriptions that are being monitored by the Datadog monitor resource.
-     *
+     * 
      * @param id the resource ID.
      * @param context The context to associate with this operation.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
@@ -137,7 +140,7 @@ public interface MonitoredSubscriptions {
 
     /**
      * Begins definition for a new MonitoredSubscriptionProperties resource.
-     *
+     * 
      * @param name resource name.
      * @return the first stage of the new MonitoredSubscriptionProperties definition.
      */