+++ sdk/datadog/azure-resourcemanager-datadog/src/main/java/com/azure/resourcemanager/datadog/models/MonitoringTagRulesProperties.java
@@ -5,42 +5,51 @@
 package com.azure.resourcemanager.datadog.models;
 
 import com.azure.core.annotation.Fluent;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.azure.json.JsonReader;
+import com.azure.json.JsonSerializable;
+import com.azure.json.JsonToken;
+import com.azure.json.JsonWriter;
+import java.io.IOException;
 
-/** Definition of the properties for a TagRules resource. */
+/**
+ * Definition of the properties for a TagRules resource.
+ */
 @Fluent
-public final class MonitoringTagRulesProperties {
+public final class MonitoringTagRulesProperties implements JsonSerializable<MonitoringTagRulesProperties> {
     /*
      * The provisioningState property.
      */
-    @JsonProperty(value = "provisioningState", access = JsonProperty.Access.WRITE_ONLY)
     private ProvisioningState provisioningState;
 
     /*
      * Set of rules for sending logs for the Monitor resource.
      */
-    @JsonProperty(value = "logRules")
     private LogRules logRules;
 
     /*
      * Set of rules for sending metrics for the Monitor resource.
      */
-    @JsonProperty(value = "metricRules")
     private MetricRules metricRules;
 
     /*
      * Configuration to enable/disable auto-muting flag
      */
-    @JsonProperty(value = "automuting")
     private Boolean automuting;
 
-    /** Creates an instance of MonitoringTagRulesProperties class. */
+    /*
+     * Configuration to enable/disable custom metrics. If enabled, custom metrics from app insights will be sent.
+     */
+    private Boolean customMetrics;
+
+    /**
+     * Creates an instance of MonitoringTagRulesProperties class.
+     */
     public MonitoringTagRulesProperties() {
     }
 
     /**
      * Get the provisioningState property: The provisioningState property.
-     *
+     * 
      * @return the provisioningState value.
      */
     public ProvisioningState provisioningState() {
@@ -48,8 +57,19 @@ public final class MonitoringTagRulesProperties {
     }
 
     /**
+     * Set the provisioningState property: The provisioningState property.
+     * 
+     * @param provisioningState the provisioningState value to set.
+     * @return the MonitoringTagRulesProperties object itself.
+     */
+    public MonitoringTagRulesProperties withProvisioningState(ProvisioningState provisioningState) {
+        this.provisioningState = provisioningState;
+        return this;
+    }
+
+    /**
      * Get the logRules property: Set of rules for sending logs for the Monitor resource.
-     *
+     * 
      * @return the logRules value.
      */
     public LogRules logRules() {
@@ -58,7 +78,7 @@ public final class MonitoringTagRulesProperties {
 
     /**
      * Set the logRules property: Set of rules for sending logs for the Monitor resource.
-     *
+     * 
      * @param logRules the logRules value to set.
      * @return the MonitoringTagRulesProperties object itself.
      */
@@ -69,7 +89,7 @@ public final class MonitoringTagRulesProperties {
 
     /**
      * Get the metricRules property: Set of rules for sending metrics for the Monitor resource.
-     *
+     * 
      * @return the metricRules value.
      */
     public MetricRules metricRules() {
@@ -78,7 +98,7 @@ public final class MonitoringTagRulesProperties {
 
     /**
      * Set the metricRules property: Set of rules for sending metrics for the Monitor resource.
-     *
+     * 
      * @param metricRules the metricRules value to set.
      * @return the MonitoringTagRulesProperties object itself.
      */
@@ -89,7 +109,7 @@ public final class MonitoringTagRulesProperties {
 
     /**
      * Get the automuting property: Configuration to enable/disable auto-muting flag.
-     *
+     * 
      * @return the automuting value.
      */
     public Boolean automuting() {
@@ -98,7 +118,7 @@ public final class MonitoringTagRulesProperties {
 
     /**
      * Set the automuting property: Configuration to enable/disable auto-muting flag.
-     *
+     * 
      * @param automuting the automuting value to set.
      * @return the MonitoringTagRulesProperties object itself.
      */
@@ -108,8 +128,30 @@ public final class MonitoringTagRulesProperties {
     }
 
     /**
+     * Get the customMetrics property: Configuration to enable/disable custom metrics. If enabled, custom metrics from
+     * app insights will be sent.
+     * 
+     * @return the customMetrics value.
+     */
+    public Boolean customMetrics() {
+        return this.customMetrics;
+    }
+
+    /**
+     * Set the customMetrics property: Configuration to enable/disable custom metrics. If enabled, custom metrics from
+     * app insights will be sent.
+     * 
+     * @param customMetrics the customMetrics value to set.
+     * @return the MonitoringTagRulesProperties object itself.
+     */
+    public MonitoringTagRulesProperties withCustomMetrics(Boolean customMetrics) {
+        this.customMetrics = customMetrics;
+        return this;
+    }
+
+    /**
      * Validates the instance.
-     *
+     * 
      * @throws IllegalArgumentException thrown if the instance is not valid.
      */
     public void validate() {
@@ -119,5 +161,55 @@ public final class MonitoringTagRulesProperties {
         if (metricRules() != null) {
             metricRules().validate();
         }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
+        jsonWriter.writeStartObject();
+        jsonWriter.writeStringField("provisioningState",
+            this.provisioningState == null ? null : this.provisioningState.toString());
+        jsonWriter.writeJsonField("logRules", this.logRules);
+        jsonWriter.writeJsonField("metricRules", this.metricRules);
+        jsonWriter.writeBooleanField("automuting", this.automuting);
+        jsonWriter.writeBooleanField("customMetrics", this.customMetrics);
+        return jsonWriter.writeEndObject();
+    }
+
+    /**
+     * Reads an instance of MonitoringTagRulesProperties from the JsonReader.
+     * 
+     * @param jsonReader The JsonReader being read.
+     * @return An instance of MonitoringTagRulesProperties if the JsonReader was pointing to an instance of it, or null
+     * if it was pointing to JSON null.
+     * @throws IOException If an error occurs while reading the MonitoringTagRulesProperties.
+     */
+    public static MonitoringTagRulesProperties fromJson(JsonReader jsonReader) throws IOException {
+        return jsonReader.readObject(reader -> {
+            MonitoringTagRulesProperties deserializedMonitoringTagRulesProperties = new MonitoringTagRulesProperties();
+            while (reader.nextToken() != JsonToken.END_OBJECT) {
+                String fieldName = reader.getFieldName();
+                reader.nextToken();
+
+                if ("provisioningState".equals(fieldName)) {
+                    deserializedMonitoringTagRulesProperties.provisioningState
+                        = ProvisioningState.fromString(reader.getString());
+                } else if ("logRules".equals(fieldName)) {
+                    deserializedMonitoringTagRulesProperties.logRules = LogRules.fromJson(reader);
+                } else if ("metricRules".equals(fieldName)) {
+                    deserializedMonitoringTagRulesProperties.metricRules = MetricRules.fromJson(reader);
+                } else if ("automuting".equals(fieldName)) {
+                    deserializedMonitoringTagRulesProperties.automuting = reader.getNullable(JsonReader::getBoolean);
+                } else if ("customMetrics".equals(fieldName)) {
+                    deserializedMonitoringTagRulesProperties.customMetrics = reader.getNullable(JsonReader::getBoolean);
+                } else {
+                    reader.skipChildren();
+                }
+            }
+
+            return deserializedMonitoringTagRulesProperties;
+        });
     }
 }